Lab_3/requirements.txt
matplotlib
numpy
mplcursors
pandas
python-docx

